Refactor ListViewAdapter in 2 Adapters for Items & Chat Menus
